+++ source/rr/src/screentext.cpp
@@ -1131,7 +1131,11 @@ void G_PrintGameQuotes(int32_t snum)
                 :
 #endif
                 (TEXT_XCENTER | (REALITY ? TEXT_N64NOPAL : 0));
+        int32_t const savedRedSplitHudDrawingView = g_redSplitHudDrawingView;
+        if (savedRedSplitHudDrawingView >= 0)
+            g_redSplitHudDrawingView = -1;
         height = gametext_(x, y, apStrings[ps->ftq], textsh(k), pal, texto(k), texta(k), quoteFlags).y + (1<<16);
+        g_redSplitHudDrawingView = savedRedSplitHudDrawingView;
     }
     while (0);